Restored lib_easy and moved controller things to own files
Maybe bloating lib_easy isn't a great idea, I've restored it and commented out all the controller bits, and moved my controller things to new files.
